% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDID.R
\name{aipwDID}
\alias{aipwDID}
\title{Chang (2020) Double-Robust semiparametric difference in differences
estimator for the ATT fit using LASSO}
\usage{
aipwDID(X, D, Y1, Y0, xfit = T)
}
\arguments{
\item{X}{NxK matrix of covariates for propensity score}

\item{D}{N-vector of treatment assignments (treatment only applies in period 2)}

\item{Y1}{N-vector of outcomes in the second period}

\item{Y0}{N-vector of outcomes in the first period}
}
\value{
ATT estimate
}
\description{
Chang (2020) Double-Robust semiparametric difference in differences
estimator for the ATT fit using LASSO
}
\references{
Abadie(2005), Chang(2020), Zhao and Sant'Anna (2021)
}
