% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDID.R
\name{balDID}
\alias{balDID}
\title{semiparametric difference in differences estimator for the ATT (balancing weights)}
\usage{
balDID(X, W, Y1, Y0)
}
\arguments{
\item{X}{NxK matrix of covariates for propensity score}

\item{W}{N-vector of treatment assignments (treatment only applies in period 2)}

\item{Y1}{N-vector of outcomes in the second period}

\item{Y0}{N-vector of outcomes in the first period}
}
\value{
ATT estimate
}
\description{
semiparametric difference in differences estimator for the ATT (balancing weights)
}
\references{
Abadie(2005)
}
