% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDID.R
\name{DID}
\alias{DID}
\title{Vanilla Difference in Differences estimation}
\usage{
DID(D, Y1, Y0)
}
\arguments{
\item{D}{N-vector of treatment assignments (treatment only applies in period 2)}

\item{Y1}{N-vector of outcomes in the second period}

\item{Y0}{N-vector of outcomes in the first period}
}
\value{
ATT estimate
}
\description{
Vanilla Difference in Differences estimation
}
\references{
Ashenfelter (1977), Snow(18xx)
}
