% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDID.R
\name{omDID}
\alias{omDID}
\title{outcome model for the ATT fit using LASSO}
\usage{
omDID(X, D, Y1, Y0, xfit = TRUE)
}
\arguments{
\item{X}{NxK matrix of covariates for propensity score}

\item{D}{N-vector of treatment assignments (treatment only applies in period 2)}

\item{Y1}{N-vector of outcomes in the second period}

\item{Y0}{N-vector of outcomes in the first period}
}
\value{
ATT estimate
}
\description{
outcome model for the ATT fit using LASSO
}
\references{
Heckman et al (1997), Abadie(2005)
}
